Remove unused files and packages
