+++ lib/descendants.js
@@ -1,96 +1,67 @@
 import {
   useState,
   createContext,
+  useContext,
   useRef,
-  useCallback,
-  useEffect,
-  useContext
+  useLayoutEffect,
+  useEffect
 } from 'react'
+import deepEqual from 'fast-deep-equal'
+
+export const useDescendantsInit = () => useState([])
 
 export const useDescendants = () => {
   const list = useRef([])
+  const [, forceUpdate] = useState()
+  const ref = useRef()
 
-  const getDOMOrder = useCallback(element => {
-    const index = list.current.findIndex(item => {
-      if (!item || !element) return false
-
-      return Boolean(
-        item.compareDocumentPosition(element) & Node.DOCUMENT_POSITION_PRECEDING
-      )
-    })
-
-    if (index === -1) {
-      return list.current.length
+  useLayoutEffect(() => {
+    if (!ref.current) {
+      return
     }
 
-    return index
-  }, [])
+    const newList = Array.from(
+      ref.current.querySelectorAll('[data-descendant]')
+    )
 
-  const remove = useCallback(element => {
-    if (!element) return
-    console.log('remove', element)
-    list.current = list.current.filter(el => el !== element)
-  }, [])
-
-  const insert = useCallback(
-    element => {
-      if (!element) {
-        return -1
-      }
-
-      const insertionIndex = getDOMOrder(element)
-      console.log('insert', element, 'at', insertionIndex)
-      list.current.splice(insertionIndex, 0, element)
-      return insertionIndex
-    },
-    [] // eslint-disable-line
-  )
+    if (
+      !deepEqual(
+        newList,
+        list.current.map(e => e.element)
+      )
+    ) {
+      list.current = newList.map(e => ({ element: e }))
+      forceUpdate({})
+    }
+  })
 
   useEffect(() => {
     return () => (list.current = [])
   }, [])
 
   return {
-    remove,
-    insert
+    listRef: ref,
+    list
   }
 }
 
 export const createDescendants = () => createContext({})
-export const useDescendant = (el, ctx) => {
-  const { insert, remove } = useContext(ctx)
-  const [count, setCount] = useState(0)
-  const [, forceUpdate] = useState()
-  const countRef = useRef(0)
+export const useDescendant = (ctx, props) => {
   const index = useRef(-1)
-  const node = useRef()
+  const ref = useRef()
+  const { list } = useContext(ctx)
 
-  // eslint-disable-next-line
-  useEffect(() => {
-    if (count !== countRef.current) {
-      // Skip
-      countRef.current = count
-      return
-    }
-
-    if (!el) {
-      forceUpdate({})
-    } else {
-      node.current = el
-      remove(el)
-      index.current = insert(el)
-      setCount(c => c + 1)
-    }
-
-    countRef.current = count
+  useLayoutEffect(() => {
+    ref.current.setAttribute('data-descendant', '')
   })
 
-  useEffect(() => {
-    return () => {
-      remove(node.current)
-    }
-    // eslint-disable-next-line
-  }, [])
+  if (ref.current) {
+    index.current = list.current.findIndex(e => e.element === ref.current)
 
-  return index.current
+    // Add any additional props to the descendant item
+    // Since it's all refs, non-memoized functions are ok!
+    list.current[index.current] = { ...list.current[index.current], ...props }
+  }
+
+  return { index: index.current, itemRef: ref }
 }
