+++ package.json
@@ -12,13 +12,8 @@
   "license": "MIT",
   "dependencies": {
     "@reach/auto-id": "^0.8.6",
-    "@reach/descendants": "^0.10.5",
     "@reach/dialog": "^0.10.5",
-    "@reach/portal": "^0.8.5",
-    "@reach/utils": "^0.10.5",
-    "body-scroll-lock": "^2.6.4",
     "classnames": "^2.2.6",
-    "country-list": "^2.2.0",
     "fast-deep-equal": "^3.1.3",
     "intersection-observer": "^0.7.0",
     "lodash.debounce": "^4.0.8",
