--- styles/inverse.module.css
@@ -1,123 +1,0 @@
-.screen {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 999;
-  --overlay: rgba(0, 0, 0, 0.6);
-  background: var(--overlay);
-  --padding: 0 20px;
-  opacity: 0;
-  transition: opacity 0.2s ease;
-}
-
-.screen.show {
-  opacity: 1;
-}
-
-.command {
-  position: relative;
-  margin: 0 auto;
-  z-index: 9999;
-  border-radius: 10px;
-  overflow: hidden;
-  top: calc(30px + var(--big-gap) + (3 * var(--gap-double)));
-  background: var(--bg);
-  box-shadow: 0px 30px 60px rgba(0, 0, 0, 0.4), 0px 0px 60px rgba(0, 0, 0, 0.5),
-    inset 0px 1px 0px rgba(0, 0, 0, 0.6);
-  width: 700px;
-
-  opacity: 0;
-  transform: scale(0.97);
-  transition: transform 0.2s ease, opacity 0.2s ease;
-}
-
-.command.show {
-  opacity: 1;
-  transform: none;
-}
-
-.command [data-command-list] {
-  margin: 0;
-  padding: 0;
-  list-style-type: none;
-}
-
-.command [data-command-input] {
-  padding: 0 var(--gap);
-  outline: none;
-  height: 60px;
-  width: 100%;
-  border: 0;
-  background: 0;
-  font-size: 1rem;
-  color: var(--fg);
-}
-
-.command [data-command-input]::placeholder {
-  color: var(--gray);
-}
-
-.command [data-command-list] {
-  margin: 0;
-  padding: 0;
-  list-style-type: none;
-}
-
-.command [data-command-item] {
-  padding: 0 var(--gap);
-  display: block;
-  outline: none;
-  font-size: 1rem;
-  cursor: pointer;
-  height: var(--height);
-  display: flex;
-  align-items: center;
-  color: var(--gray);
-  min-height: 60px;
-}
-
-.command [data-command-item][aria-selected] {
-  background: var(--lightest-gray);
-  color: var(--fg);
-}
-
-.top {
-  display: flex;
-  align-items: center;
-  padding: var(--padding);
-}
-
-.back {
-  background: none;
-  color: #fff;
-}
-
-.label {
-  display: flex;
-  align-items: center;
-  min-height: 5px;
-  color: var(--gray);
-  padding: var(--padding);
-  padding-top: 2px;
-  padding-bottom: 2px;
-  font-size: 0.875rem;
-  background: var(--lighter-gray);
-}
-
-.container {
-  max-height: 300px;
-  overflow-y: auto;
-  -webkit-overflow-scrolling: touch;
-  transition: height 0.1s ease;
-  will-change: height;
-}
-
-.container {
-  border-top: 1px solid var(--lighter-gray);
-}
-
-.container.empty {
-  border-top: none;
-}
